☀️ MERGE : main으로 merge 합니다

+++ src/main/java/com/example/devlife/controller/CommentController.java
@@ -1,9 +1,12 @@
 package com.example.devlife.controller;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -21,42 +24,56 @@ import com.example.devlife.entity.Comment;
 import com.example.devlife.entity.User;
 import com.example.devlife.service.comment.CommentService;
 
+import jakarta.transaction.Transactional;
+import lombok.RequiredArgsConstructor;
+
 @Controller
+@RequiredArgsConstructor
 public class CommentController {
 	private final CommentService commentService;
 
-	public CommentController(CommentService commentService) {
-		this.commentService = commentService;
-	}
-
 	@PostMapping("/api/comment/{boardId}")
-	public ResponseEntity<Comment> addComment(@PathVariable(value = "boardId") Long boardId,
-		@RequestBody AddCommentRequest request) {
-		//request - content, parentId, depth, sequence
-		//TODO 0. User 로그인 정보 확인
-		User user = null;
+	@PreAuthorize("isAuthenticated()")
+	public ResponseEntity<Comment> addComment(
+			@AuthenticationPrincipal(expression = "#this == 'anonymousUser' ? null : account") User user,
+			@PathVariable(value = "boardId") Long boardId,
+			@RequestBody AddCommentRequest request) {
 		Comment savedComment = commentService.save(user, boardId, request);
 		return ResponseEntity.status(HttpStatus.CREATED).body(savedComment);
 	}
 
 	@GetMapping("/api/comment/{boardId}")
-	public ResponseEntity<List<Comment>> readComment(@PathVariable(value = "boardId") Long boardId) {
-		//TODO 1. User 로그인 정보 확인
+	@PreAuthorize("isAuthenticated()")
+	public ResponseEntity<List<Comment>> readComment(
+			@PathVariable(value = "boardId") Long boardId) {
 		List<Comment> comments = commentService.getComments(boardId);
+		
 		return ResponseEntity.ok().body(comments);
 	}
 
 	@PatchMapping("/api/comment/{commentId}")
-	public ResponseEntity<Comment> updateComment(@PathVariable(value = "commentId") Long commentId, @RequestBody
-		UpdateCommentRequest request) {
-		//TODO 1. User 로그인 정보 확인
+	@PreAuthorize("isAuthenticated()")
+	public ResponseEntity<Comment> updateComment(
+			@AuthenticationPrincipal(expression = "#this == 'anonymousUser' ? null : account") User user,
+			@PathVariable(value = "commentId") Long commentId,
+			@RequestBody UpdateCommentRequest request) {
+		Comment originalComment = commentService.getComment(commentId);
+		if (originalComment.getUser().getId() != user.getId()) {
+			return ResponseEntity.badRequest().body(null);
+		}
 		Comment updatedComment = commentService.updateComment(commentId, request);
 		return ResponseEntity.ok().body(updatedComment);
 	}
 
 	@DeleteMapping("/api/comment/{commentId}")
-	public ResponseEntity<Void> deleteComment(@PathVariable(value = "commentId") Long commentId) {
-		//TODO 1. User 로그인 정보 확인
+	@PreAuthorize("isAuthenticated()")
+	public ResponseEntity<Void> deleteComment(
+			@AuthenticationPrincipal(expression = "#this == 'anonymousUser' ? null : account") User user,
+			@PathVariable(value = "commentId") Long commentId) {
+		Comment originalComment = commentService.getComment(commentId);
+		if (originalComment.getUser().getId() != user.getId()) {
+			return ResponseEntity.badRequest().body(null);
+		}
 		commentService.deleteComment(commentId);
 		return ResponseEntity.ok().body(null);
 	}
